feat(*): port process interface to new io
