+++ library/tools/smt2/solvers/z3.lean
@@ -3,22 +3,31 @@ import system.process
 import system.read
 import data.buffer
 
-meta structure z3_instance :=
-    (process : process.child)
+open process
 
-meta def z3_instance.start : io z3_instance := do
-    builder <- process.builder.new "z3",
-    builder^.arg "-smt2",
-    builder^.arg "-in",
-    builder^.stdin process.stdio.piped,
-    builder^.stdout process.stdio.piped,
-    z3_instance.mk <$> builder^.spawn
+structure z3_instance [io.interface] : Type :=
+    (process : child io.handle)
 
-meta def z3_instance.raw (z3 : z3_instance) (cmd : string) : io.result string := do
-    stdin <- z3^.process^.stdin,
-    stdout <- z3^.process^.stdout,
+meta def spawn [ioi : io.interface] : process → io (child io.handle) :=
+    io.interface.process^.spawn
+
+meta def z3_instance.start  [io.interface] : io z3_instance := do
+    let proc : process := {
+       cmd := "z3",
+       args := ["-smt2", "-in"],
+       stdin := stdio.piped,
+       stdout := stdio.piped
+    },
+    z3_instance.mk <$> spawn proc
+
+meta def z3_instance.raw [io.interface] (z3 : z3_instance) (cmd : string) : io string := do
+    let z3_stdin := z3^.process^.stdin,
+    let z3_stdout := z3^.process^.stdout,
     -- do we buffer the writes here?
-    let cs := cmd^.reverse^.to_buffer^.to_array,
-    process.pipe.write stdin cs,
-    process.pipe.write stdin (mk_array 1 (char.of_nat 0)),
-    read_to_string stdout
+    let cs := cmd^.reverse^.to_buffer,
+    io.fs.write z3_stdin cs,
+    res <- io.fs.read z3_stdout 1024,
+    return res^.to_string
+
+    -- process.pipe.write stdin (mk_array 1 (char.of_nat 0)),
+    -- read_to_string stdout
