+++ src/library/CMakeLists.txt
@@ -19,4 +19,4 @@ add_library(library OBJECT deep_copy.cpp expr_lt.cpp io_state.cpp
   mt_task_queue.cpp st_task_queue.cpp
   library_task_builder.cpp
   messages.cpp message_builder.cpp module_mgr.cpp comp_val.cpp
-  documentation.cpp check.cpp arith_instance.cpp parray.cpp)
+  documentation.cpp check.cpp arith_instance.cpp parray.cpp process.cpp pipe.cpp)
