+++ src/library/handle.cpp
@@ -1,0 +1,32 @@
+/*
+Copyright (c) 2017 Microsoft Corporation. All rights reserved.
+Released under Apache 2.0 license as described in the file LICENSE.
+
+Author:  Leonardo de Moura & Jared Roesch
+*/
+#include <string>
+#include <fstream>
+#include <iostream>
+#include <iomanip>
+#include <utility>
+#include <unistd.h>
+#if defined(LEAN_WINDOWS) && !defined(LEAN_CYGWIN)
+#include <windows.h>
+#include <tchar.h>
+#include <stdio.h>
+#include <strsafe.h>
+
+#define BUFSIZE 4096
+#else
+#include <sys/wait.h>
+#endif
+#include "library/process.h"
+#include "util/buffer.h"
+#include "library/pipe.h"
+
+namespace lean {
+}
+
+#endif
+
+}
