+++ src/library/native_compiler/cpp_compiler.cpp
@@ -5,7 +5,7 @@ Released under Apache 2.0 license as described in the file LICENSE.
 Author: Jared Roesch
 */
 #include <string>
-#include "util/process.h"
+#include "library/process.h"
 #include "library/native_compiler/cpp_compiler.h"
 
 namespace lean {
