+++ src/library/pipe.h
