+++ src/library/vm/init_module.cpp
@@ -25,7 +25,6 @@ Author: Leonardo de Moura
 #include "library/vm/vm_parser.h"
 #include "library/vm/vm_array.h"
 #include "library/vm/vm_process.h"
-#include "library/vm/vm_file.h"
 
 namespace lean {
 void initialize_vm_core_module() {
@@ -50,11 +49,9 @@ void initialize_vm_core_module() {
     initialize_vm_parser();
     initialize_vm_array();
     initialize_vm_process();
-    initialize_vm_file();
 }
 
 void finalize_vm_core_module() {
-    finalize_vm_file();
     finalize_vm_process();
     finalize_vm_array();
     finalize_vm_parser();
