+++ src/library/vm/vm_io.h
@@ -7,6 +7,8 @@ Author: Leonardo de Moura
 #pragma once
 #include <string>
 #include "library/vm/vm.h"
+#include "library/handle.h"
+
 namespace lean {
 vm_obj mk_io_result(vm_obj const & r);
 vm_obj mk_io_interface();
@@ -18,6 +20,11 @@ optional<vm_obj> is_io_result(vm_obj const & o);
 optional<vm_obj> is_io_error(vm_obj const & o);
 /* Convert an io.error object into a string */
 std::string io_error_to_string(vm_obj const & o);
+
+bool is_handle(vm_obj const & o);
+handle_ref const & to_handle(vm_obj const & o);
+vm_obj to_obj(handle_ref const & h);
+
 void initialize_vm_io();
 void finalize_vm_io();
 }
