+++ src/util/CMakeLists.txt
@@ -6,5 +6,4 @@ add_library(util OBJECT debug.cpp name.cpp name_set.cpp fresh_name.cpp
   utf8.cpp name_map.cpp list_fn.cpp null_ostream.cpp file_lock.cpp
   timer.cpp task.cpp task_builder.cpp cancellable.cpp
   log_tree.cpp
-  small_object_allocator.cpp subscripted_name_set.cpp process.cpp
-  parser_exception.cpp pipe.cpp)
+  small_object_allocator.cpp subscripted_name_set.cpp parser_exception.cpp)
